pyproject.toml
[project]
requires-python = ">=3.7,<4.0"

[tool.poetry]
name = "idf-component-manager"
version = "1.5.2"
description = "Espressif IDF Component Manager"
authors = []
license = "Apache2"
readme = "README.md"
packages = [
    { include = "idf_component_manager" },
    { include = "idf_component_tools" },
]

[tool.poetry.dependencies]
cachecontrol = { version = '*', extras = ["filecache"] }
click = '*'
colorama = '*'
packaging = "*"
pyparsing = '*'
python = ">=3.7,<4.0"
pyyaml = "*"
requests = "*"
requests-file = "*"
requests-toolbelt = "*"
schema = "*"
tqdm = '*'
urllib3 = "<2"

[tool.poetry.group.docs.dependencies]
sphinx_collapse = "==0.1.2"
sphinx-click = "==4.4.0"
sphinx-copybutton = "==0.5.1"
sphinx-rtd-theme = "==1.1.1"
sphinx-tabs = "==3.4.0"

[tool.poetry.group.dev.dependencies]
black = { version = "==23.7.0", python = ">=3.8,<4.0" }
comment-parser = "*"
commitizen = { version = "^3.13.0", python = ">=3.8,<4.0" }
coverage = '*'
deptry = { version = ">=0.12.0", python = ">=3.9,<4.0" }
flake8 = '*'
flake8-pyproject = "*"
flake8-quotes = '*'
jinja2 = "*"
jsonschema = "*"
mypy = "*"
pexpect = "*"
pre-commit = '*'
pytest = '*'
pytest-cov = '*'
pytest-mock = '*'
pytest-rerunfailures = "*"
requests-mock = "^1.11.0"
rope = '*'
setuptools = "*"
thefuzz = "*"
twine = "*"
types-pyyaml = "^6.0.12.10"
types-requests = "^2.31.0.1"
unify = '*'
vcrpy = '<5'
wheel = "*"

[tool.poetry.scripts]
compote = { callable = "idf_component_manager.cli:safe_cli" }

[tool.isort]
include_trailing_comma = true
use_parentheses = true
ensure_newline_before_comments = true
line_length = 100
indent = '    '
multi_line_output = 3
force_grid_wrap = 0

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.7"

[[tool.mypy.overrides]]
module = "idf_component_manager.version_solver.mixology.*"
ignore_errors = true

[tool.deptry]
extend_exclude = [
    "docs",
    "integration_tests",
    "idf_component_tools/file_cache.py", # Tries a number of ways to detect paths on Windows
]

[tool.deptry.per_rule_ignores]
DEP001 = ["urlparse", "idf_py_actions", "functools32", "httplib"]
DEP002 = ["colorama", "urllib3"]

[tool.deptry.package_module_name_map]
pyyaml = ["yaml"]

[tool.commitizen]
name = "cz_conventional_commits"
tag_format = "v$version"
version_scheme = "semver"
version_provider = "poetry"
version_files = ["idf_component_tools/__version__.py:__version__"]
update_changelog_on_bump = true
bump_message = "change: bump version from v$current_version to v$new_version"

[tool.ruff]
line-length = 100
target-version = "py37"

[tool.ruff.lint]
extend-select = [
    "S", # flake8-bandit
    "I",  # isort
    "ERA",  # eradicate
    "ICN",  # flake8-import-conventions
]

[tool.ruff.format]
quote-style = "single"
docstring-code-format = true

[tool.ruff.lint.flake8-import-conventions]
banned-from = ["typing"]

[tool.ruff.lint.flake8-import-conventions.extend-aliases]
"typing" = "t"

[tool.ruff.lint.per-file-ignores]
"setup.py" = [
    "S",
]

"ci/*" = [
    "S",
]

"integration_tests/*" = [
    "S",
    "E712",  # comparison to False/True
]

"tests/*" = [
    "S",
    "E712",  # comparison to False/True
]

"idf_component_tools/semver/base.py" = [
    "S101",  # use of assert
]

"idf_component_manager/version_solver/mixology/*" = [
    "S101",  # use of assert
]
